fusion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.6
aiohttp==3.9.1
pyyaml==6.0.1
pillow==10.1.0

